+++ example/pom.xml
@@ -33,5 +33,10 @@
             <groupId>org.eclipse.jdt</groupId>
             <artifactId>org.eclipse.jdt.annotation</artifactId>
         </dependency>
+        <dependency>
+            <groupId>be.kuleuven.cs</groupId>
+            <artifactId>gridflex-project</artifactId>
+            <version>0.1.4-SNAPSHOT</version>
+        </dependency>
     </dependencies>
 </project>
