+++ example/src/test/java/be/kuleuven/cs/gametheory/GameScenarioTest.java
@@ -1,10 +1,11 @@
 package be.kuleuven.cs.gametheory;
 
-import be.kuleuven.cs.gridflex.domain.aggregation.Aggregator;
-import be.kuleuven.cs.gridflex.domain.site.Site;
-import be.kuleuven.cs.gridflex.experimentation.DefaultGameConfigurator;
 import be.kuleuven.cs.gametheory.standalone.Game;
 import be.kuleuven.cs.gametheory.standalone.GameDirector;
+import be.kuleuven.cs.gridflex.domain.aggregation.Aggregator;
+import be.kuleuven.cs.gridflex.domain.site.Site;
+import be.kuleuven.cs.gridflex.experimentation.games.DefaultGameConfigurator;
+import be.kuleuven.cs.gridflex.experimentation.util.GameResultWriter;
 import org.junit.Before;
 import org.junit.Test;
