+++ project/pom.xml
@@ -27,11 +27,6 @@
         </dependency>
         <dependency>
             <groupId>be.kuleuven.cs</groupId>
-            <artifactId>gridflex-example</artifactId>
-            <version>0.1.4-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>be.kuleuven.cs</groupId>
             <artifactId>gridflex-persistence</artifactId>
             <version>0.1.4-SNAPSHOT</version>
         </dependency>
