+++ project/src/main/java/be/kuleuven/cs/gridflex/experimentation/games/package-info.java
@@ -4,6 +4,6 @@
  * @author Kristof Coninx (kristof.coninx AT cs.kuleuven.be)
  */
 @NonNullByDefault
-package be.kuleuven.cs.gridflex.experimentation;
+package be.kuleuven.cs.gridflex.experimentation.games;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
