+++ project/src/main/java/be/kuleuven/cs/gridflex/experimentation/tosg/wgmf/WgmfGameRunner.java
@@ -85,6 +85,6 @@ public class WgmfGameRunner extends AbstractWgmfGameRunner {
         }
         logger.warn("Dynamics equation params: {}", director.getDynamicEquationArguments());
         logger.warn("Payoff table: \n{}",
-                director.getFormattedResults().getFormattedResultString());
+                director.getFormattedResults());
     }
 }
